backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Django backend for The Grumpy Tracker"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "camelot-py[cv]>=1.0.0",
    "django>=5.2.1",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "pdfplumber>=0.11.6",
    "psycopg[binary]>=3.2.9",
    "python-decouple>=3.8",
]
